+++ src/main/java/com/devicesus/dto/DeviceDto.java
@@ -15,5 +15,5 @@ public class DeviceDto{
     private String description;
     private String address;
     // add IP address validator (IPv6 or IPv4)
-    private Long maxHourlyEnergyConsumption;
+    private Double maxHourlyEnergyConsumption;
 }
